requirements.txt
Hydrogram~=0.2.0
python-dotenv~=1.0.1
redis~=5.2.0
TgCrypto~=1.2.5
yt-dlp~=2024.10.22
